+++ changelog.md
@@ -1,6 +1,7 @@
 ## v1.1.0
 - Port to Geode v5 / Geometry Dash 2.2081
 - Added bugfix for horizontal scrolling on Windows
+- Editor camera scroll support 
 - Improve compatibility with most mods by checking for active scroll layers
 
 ## v1.0.1